move contexts out to own folder, update imports for UserContext
