+++ src/contexts/UserContext.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { ZetkinUser } from '../../types/zetkin';
+import { ZetkinUser } from '../types/zetkin';
 
 const UserContext = React.createContext<ZetkinUser | null>(null);
